Fixed issue with QRC file registration introduced in Qt 5.15.
Introduced a new qrc path: ":/esri.com/import". This has changed
the invocation args of registerComponents. QML toolkit now must call
"registerComponents."

Fixed issue with NorthArrow image path in widgets.
