+++ uitools/README.md
@@ -35,20 +35,33 @@ get added to QtCreator when you install the ArcGIS Runime SDK for Qt.
 include(path/to/toolkitqml.pri)
 ```
 
+- in `main.cpp` add a line to import the toolkit registration function.
+
+```cpp
+#include "Esri/ArcGISRuntime/Toolkit/register.h"
+```
+
+- in `main.cpp` (inside the main function) call the registration function.
+
+```cpp
+QQmlApplicationEngine engine; // The engine driving your QML application.
+Esri::ArcGISRuntime::Toolkit::registerComponents(engine);
+```
+
 #### Using a tool from the toolkit (toolkitqml.pri)
 
 Once you have successfully imported the toolkit, you can access individual tools
 in your own QML files.
 
 - add an import statement for the toolkit:
-  `import Esri.ArcGISRuntime.Toolkit 100.9`
+  `import Esri.ArcGISRuntime.Toolkit 100.10`
 - declare the tool you wish to use. Generally you will also have to supply the
   `GeoView` etc. you wish the tool to work with. For example, to add a
   `NorthArrow`:
 
 ```qml
-import Esri.ArcGISRuntime 100.9
-import Esri.ArcGISRuntime.Toolkit 100.9
+import Esri.ArcGISRuntime 100.10
+import Esri.ArcGISRuntime.Toolkit 100.10
 
 MapView {
     id: mapView
@@ -78,7 +91,7 @@ MapView {
 A good way to start using the toolkit is to use one of the template apps which
 get added to QtCreator when you install the ArcGIS Runime SDK for Qt.
 
-- In QtCreator choose `File/New file or project/ArcGIS/ArcGIS Runtime 100.9 Qt Quick C++ App`
+- In QtCreator choose `File/New file or project/ArcGIS/ArcGIS Runtime 100.10 Qt Quick C++ App`
 - choose settings to match the platform etc. you are building for
 - in the `.pro` file of your new app, add a line to add the library for your QML
   plugin - for example:
@@ -100,7 +113,7 @@ include(path/to/toolkitcpp.pri)
 
 ```cpp
 QQmlApplicationEngine engine; // The engine driving your QML application.
-Esri::ArcGISRuntime::Toolkit::registerComponents(&engine);
+Esri::ArcGISRuntime::Toolkit::registerComponents(engine);
 ```
 
 #### Using a tool from the toolkit (toolkitcpp.pri)
@@ -109,12 +122,12 @@ Once you have successfully imported the toolkit, you can access individual tools
 in your own QML files.
 
 - add an import statement for the toolkit:
-`import Esri.ArcGISRuntime.Toolkit 100.9`
+`import Esri.ArcGISRuntime.Toolkit 100.10`
 - declare the tool you wish to use. Generally you will also have to supply the
   `GeoView` etc. you wish the tool to work with. For example, to add a `NorthArrow`:
 
 ```qml
-import Esri.ArcGISRuntime.Toolkit 100.9
+import Esri.ArcGISRuntime.Toolkit 100.10
 
 MapView {
     id: mapView
@@ -140,7 +153,7 @@ MapView {
 A good way to start using the toolkit is to use one of the template apps which
 get added to QtCreator when you install the ArcGIS Runtime SDK for Qt.
 
-- In QtCreator choose `File/New file or project/ArcGIS/ArcGIS Runtime 100.9 Qt Widgets App`
+- In QtCreator choose `File/New file or project/ArcGIS/ArcGIS Runtime 100.10 Qt Widgets App`
 - choose settings to match the platform etc. you are building for
 - in the `.pro` file of your new app, add a line to add the library for your QML
   plugin - for example:
