+++ uitools/import/Esri/ArcGISRuntime/Toolkit/esri_arcgisruntime_toolkit_view.qrc
@@ -1,6 +1,6 @@
 <RCC>
   <!-- This section contains all view components (shared by C++ and QML APIs) -->
-  <qresource prefix="/qt-project.org/imports/Esri/ArcGISRuntime/Toolkit/">
+  <qresource prefix="/esri.com/imports/Esri/ArcGISRuntime/Toolkit/">
     <file>AuthenticationView.qml</file>
     <file>Callout.qml</file>
     <file>ClientCertificateView.qml</file>
