+++ uitools/widgets/Esri/ArcGISRuntime/Toolkit/NorthArrow.cpp
@@ -52,7 +52,7 @@ NorthArrow::NorthArrow(QWidget* parent) :
   QLabel(parent),
   m_controller(new NorthArrowController(this))
 {
-  m_image = QPixmap(":/Esri/ArcGISRuntime/Tookit/images/compass.png");
+  m_image = QPixmap(":/esri.com/Esri/ArcGISRuntime/Tookit/images/compass.png");
 
   if (!m_image.isNull())
   {
